Matt's Stuff/spifi-api-master/requirements.txt
Django==2.0.12
djangorestframework==3.9.1